Handle login with google
We use a custom-session-store helper because passport library doesn't allow us to pass custom state value to google url. We need to pass the redirect url for example and potentially other data through it. This helper function will let us add custom state passing by the session in the middleware, then the function will take what there is into the session and store it into the state value to pass it to the google redirection url.